feat: enhance JellyUserDetail and PlexUserDetail components with new modal functionality and provisioning options
- Added LoaderCircle and X icons for loading states and modal close actions.
- Refactored user detail interfaces to include provisioning options and action responses.
- Implemented modal dialogs for renewing, deleting, changing password, and updating notes.
- Improved user experience with estimated cost calculations for renewals.
- Added success and error notifications for user actions.
- Updated UI components to use a consistent styling approach.

+++ backend/models.py
@@ -10,7 +10,7 @@ class Reseller(Base):
     username = Column(String(50), unique=True, nullable=False)
     master = Column(Integer, nullable=True)
     password = Column(Text, nullable=False)
-    credito = Column(Integer, default=0)
+    credito = Column(Float, default=0)
     idtelegram = Column(BigInteger, nullable=True)
     ruolo = Column(String(20), nullable=False, default="reseller")
